Refactor layer hooks
