+++ packages/map/src/layer/hooks.jsx
@@ -1,5 +1,6 @@
 import { useEffect, useRef } from 'react';
 import { useMapContext } from '../map/hooks';
+import { applyPaint, resetPaint, setCursor } from './utils';
 
 export const useMapRef = () => {
   const { map } = useMapContext();
@@ -30,4 +31,58 @@ export const useMapEffect = (fn, deps) => {
       if (teardown) teardown();
     };
   }, effectDeps);
+};
+
+export const useLayer = (layerId, { type, paint, geoJsonSource }) => {
+  useMapEffect((map) => {
+    map.addLayer({
+      type,
+      paint,
+      source: geoJsonSource,
+      id: layerId,
+    });
+
+    return () => {
+      map.removeLayer(layerId);
+      map.removeSource(layerId);
+    };
+  }, []);
+};
+
+export const useLayerSource = (layerId, geoJsonSource) => {
+  useMapEffect((map) => {
+    if (!geoJsonSource) return;
+
+    const source = map.getSource(layerId);
+    source.setData(geoJsonSource.data);
+  }, [geoJsonSource]);
+};
+
+export const useLayerClick = (layerId, onClick) => {
+  useMapEffect((map) => {
+    if (!onClick) return;
+
+    const setPointerCursor = setCursor.bind(undefined, map, 'pointer');
+    const resetCursor = setCursor.bind(undefined, map, null);
+
+    map.on('click', layerId, onClick);
+    map.on('mousemove', layerId, setPointerCursor);
+    map.on('mouseleave', layerId, resetCursor);
+
+    return () => {
+      map.off('click', layerId, onClick);
+      map.off('mousemove', layerId, setPointerCursor);
+      map.off('mouseleave', layerId, resetCursor);
+    };
+  }, [onClick]);
+};
+
+export const useLayerPaint = (layerId, paint) => {
+  useMapEffect((map) => {
+    applyPaint(map, layerId, paint);
+
+    return () => {
+      resetPaint(map, layerId, paint);
+    };
+  }, [paint]);
 };
